+++ CHANGELOG.rst
@@ -1,3 +1,13 @@
+Dev
+***
+
+Fixed
+.....
+
+* Fix #122: Add support for arrays of number of trials in a mixture of
+  multinomials.
+
+
 Version 0.5.21 (2021-03-04)
 +++++++++++++++++++++++++++